refactor func to calculate timeframe to fetch events for
